moving title for a case to formhelper
