+++ apps/judicial-system/web/src/utils/formHelper.ts
@@ -1,16 +1,24 @@
 import compareAsc from 'date-fns/compareAsc'
 
 import { formatDate } from '@island.is/judicial-system/formatters'
-import type { CaseFile } from '@island.is/judicial-system/types'
+import {
+  CaseDecision,
+  CaseFile,
+  CaseState,
+  CaseType,
+  isInvestigationCase,
+} from '@island.is/judicial-system/types'
 import {
   TempCase as Case,
   TempUpdateCase as UpdateCase,
 } from '@island.is/judicial-system-web/src/types'
 import * as constants from '@island.is/judicial-system/consts'
+import { signedVerdictOverview as m } from '@island.is/judicial-system-web/messages'
 
 import { padTimeWithZero, parseTime, replaceTabs } from './formatters'
 import { TUploadFile } from './hooks'
 import * as validations from './validate'
+import { IntlShape } from 'react-intl'
 
 export const removeTabsValidateAndSet = (
   field: keyof UpdateCase,
@@ -360,3 +368,36 @@ export const mapCaseFileToUploadFile = (file: CaseFile): TUploadFile => ({
   category: file.category,
   policeCaseNumber: file.policeCaseNumber,
 })
+
+export const titleForCase = (
+  formatMessage: IntlShape['formatMessage'],
+  theCase: Case,
+) => {
+  const isTravelBan =
+    theCase.decision === CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN ||
+    theCase.type === CaseType.TRAVEL_BAN
+
+  if (theCase.state === CaseState.REJECTED) {
+    if (isInvestigationCase(theCase.type)) {
+      return 'Kröfu um rannsóknarheimild hafnað'
+    } else {
+      return 'Kröfu hafnað'
+    }
+  }
+
+  if (theCase.state === CaseState.DISMISSED) {
+    return formatMessage(m.dismissedTitle)
+  }
+
+  if (theCase.isValidToDateInThePast) {
+    return formatMessage(m.validToDateInThePast, {
+      caseType: isTravelBan ? CaseType.TRAVEL_BAN : theCase.type,
+    })
+  }
+
+  return isInvestigationCase(theCase.type)
+    ? formatMessage(m.investigationAccepted)
+    : formatMessage(m.restrictionActive, {
+        caseType: isTravelBan ? CaseType.TRAVEL_BAN : theCase.type,
+      })
+}
